Fixing vectors with integers by longs to adequate as indexes

+++ Projects/AcoplamentoH1Hdiv/AcoplamentoH1Hdiv.cpp
@@ -272,8 +272,8 @@ int main()
 		
 		TPZAdmChunkVector<TPZCompEl *> elvec = comp->ElementVec();
 		
-		int iel;
-		int nel = elvec.NElements();
+		long iel;
+		long nel = elvec.NElements();
 		for (iel=0; iel<nel; iel++) {
 				TPZCompEl *cel = comp->ElementVec()[iel];
 				if(!cel) continue;
@@ -338,9 +338,9 @@ TPZGeoMesh *MalhaGeoGen(int h, REAL &anglo)
 			
 		gMesh->NodeVec().Resize(Qnodes);
 		
-		TPZVec <int> TopolQuad(4);
-		TPZVec <int> TopolLine(2);
-		TPZVec <int> TopolPoint(1);
+		TPZVec <long> TopolQuad(4);
+		TPZVec <long> TopolLine(2);
+		TPZVec <long> TopolPoint(1);
 		
 		TPZVec<TPZGeoNode> Node(Qnodes);
 		
@@ -391,7 +391,7 @@ TPZGeoMesh *MalhaGeoGen(int h, REAL &anglo)
 		int nod;
 		TPZVec<REAL> coord(dim);
 		for(nod=0; nod<nnode; nod++) {
-				int nodind = gMesh->NodeVec().AllocateNewElement();
+				long nodind = gMesh->NodeVec().AllocateNewElement();
 				
 				for(int d = 0; d < dim; d++)
 				{
@@ -406,18 +406,18 @@ TPZGeoMesh *MalhaGeoGen(int h, REAL &anglo)
 			
 	//	int index;
 		
-		TPZVec<int> nodind1(4);
+		TPZVec<long> nodind1(4);
 		nodind1[0] = 0;
 		nodind1[1] = 1;
 		nodind1[2] = 4;
 		nodind1[3] = 5;
 		
-		int id=0;
+		long id=0;
 		
 		new TPZGeoElRefPattern< pzgeom::TPZGeoQuad> (id,nodind1, quadmat1,*gMesh);
 		id ++;
 		
-		TPZVec<int> nodind2(4);
+		TPZVec<long> nodind2(4);
 		nodind2[0] = 1;
 		nodind2[1] = 2;
 		nodind2[2] = 3;
@@ -426,7 +426,7 @@ TPZGeoMesh *MalhaGeoGen(int h, REAL &anglo)
 		id ++;
 		
 		//-----------------------------------elemento de interface---------------------------
-		TPZVec<int> nodind3(2);
+		TPZVec<long> nodind3(2);
 		
 		nodind3[0]=1;
 		nodind3[1]=4;
@@ -483,8 +483,8 @@ TPZGeoMesh *MalhaGeoGen(int h, REAL &anglo)
 		//	Refinamento uniforme
 		for(int ref = 0; ref < h; ref++){// h indica o numero de refinamentos
 				TPZVec<TPZGeoEl *> filhos;
-				int n = gMesh->NElements();
-				for(int i = 0; i < n; i++){
+				long n = gMesh->NElements();
+				for(long i = 0; i < n; i++){
 						TPZGeoEl * gel = gMesh->ElementVec()[i];
 																			
 						if(!gel->HasSubElement())
@@ -510,7 +510,7 @@ TPZFMatrix<REAL> MatrixR(REAL ang)
 		
 		return r;
 }
-TPZCompEl *CreateInterfaceEl(TPZGeoEl *gel,TPZCompMesh &mesh,int &index) {
+TPZCompEl *CreateInterfaceEl(TPZGeoEl *gel,TPZCompMesh &mesh,long &index) {
 		if(!gel->Reference() && gel->NumInterfaces() == 0)
 				return new TPZInterfaceElement(mesh,gel,index);
 		
@@ -642,7 +642,7 @@ TPZCompMesh *MalhaCompGen(TPZGeoMesh * gMesh, int porder)
 	//	comp->CleanUpUnconnectedNodes();
 		
 		//AQUI: Criar elemento de interface
-		for(int el = 0; el < comp->ElementVec().NElements(); el++)
+		for(long el = 0; el < comp->ElementVec().NElements(); el++)
 		{
 				TPZCompEl * compEl = comp->ElementVec()[el];
 				if(!compEl) continue;
@@ -654,11 +654,11 @@ TPZCompMesh *MalhaCompGen(TPZGeoMesh * gMesh, int porder)
 				
 		}
 		
-		for(int el = 0; el < comp->ElementVec().NElements(); el++)
+		for(long el = 0; el < comp->ElementVec().NElements(); el++)
 		{
 				TPZCompEl * compEl = comp->ElementVec()[el];
 				if(!compEl) continue;
-				int index = compEl ->Index();
+				long index = compEl ->Index();
 				if((compEl) && (compEl->Dimension() == 2) && (compEl->Reference()->MaterialId() == quadmat2))
 				{
 						TPZInterpolationSpace * InterpEl = dynamic_cast<TPZInterpolationSpace *>(comp->ElementVec()[index]);
@@ -719,8 +719,8 @@ void SolveLU ( TPZAnalysis &an ){
 
 void PrintInterface(TPZCompMesh *malha)
 {
-		int el;
-		const int nelem = malha->NElements();
+		long el;
+		const long nelem = malha->NElements();
 		
 		TPZInterfaceElement *face;
 		
@@ -733,7 +733,7 @@ void PrintInterface(TPZCompMesh *malha)
 						
 						face = dynamic_cast<TPZInterfaceElement *> (Cel);
 						if(face){
-								int index = Cel->Index();
+								long index = Cel->Index();
 								cout<<endl;
 								cout << "Ok ... O elemento de Index "<< index <<" e matId = "<< matId<<" eh de  interface"<<endl;
 								Cel->Reference()->Print();
@@ -766,11 +766,11 @@ void SaddlePermute(TPZCompMesh * cmesh){
 		//		DebugStop();
 		//	}
 		
-		int jperm=0;
-		int nel=cmesh->ElementVec().NElements();
-		for (int jel=0; jel<nel; jel++) {
+		long jperm=0;
+		long nel=cmesh->ElementVec().NElements();
+		for (long jel=0; jel<nel; jel++) {
 				
-				for (int ip=0; ip<permute.NElements(); ip++) {
+				for (long ip=0; ip<permute.NElements(); ip++) {
 						permute[ip]=ip;
 				}
 				
@@ -778,16 +778,16 @@ void SaddlePermute(TPZCompMesh * cmesh){
 				
 				
 				//	int idtroca=0;
-				int eqmax=0;
+				long eqmax=0;
 				if(!elvec)continue;
-				int ncon=elvec->NConnects();
+				long ncon=elvec->NConnects();
 								
 				
 				//	if(ncon==1) continue;
-				int eqpress=elvec->Connect(ncon-1).SequenceNumber();
-				for (int icon=0; icon< ncon-1; icon++) {
+				long eqpress=elvec->Connect(ncon-1).SequenceNumber();
+				for (long icon=0; icon< ncon-1; icon++) {
 						TPZConnect &coel=elvec->Connect(icon);
-						int eqflux=coel.SequenceNumber();
+						long eqflux=coel.SequenceNumber();
 						if (eqflux >= numinternalconnects) {
 								continue;
 						}
@@ -859,31 +859,31 @@ TPZGeoMesh * MalhaGeoT(const int h){//malha triangulo
 		
 		//int index;
 		
-		TPZVec<int> nodind1(3);
+		TPZVec<long> nodind1(3);
 		nodind1[0] = 0;
 		nodind1[1] = 1;
 		nodind1[2] = 5;
 				
-		int id=0;
+		long id=0;
 		
 		new TPZGeoElRefPattern< pzgeom::TPZGeoTriangle> (id,nodind1, quadmat1,*gmesh);
 		id ++;
 		
-		TPZVec<int> nodind2(3);
+		TPZVec<long> nodind2(3);
 		nodind2[0] = 4;
 		nodind2[1] = 5;
 		nodind2[2] = 1;
 				new TPZGeoElRefPattern< pzgeom::TPZGeoTriangle > (id,nodind2, quadmat1,*gmesh);
 		id ++;
 		
-		TPZVec<int> nodind3(3);
+		TPZVec<long> nodind3(3);
 		nodind3[0] = 1;
 		nodind3[1] = 2;
 		nodind3[2] = 4;
 		new TPZGeoElRefPattern< pzgeom::TPZGeoTriangle > (id,nodind3, quadmat2,*gmesh);
 		id ++;
 		
-		TPZVec<int> nodind4(3);
+		TPZVec<long> nodind4(3);
 		nodind4[0] = 3;
 		nodind4[1] = 4;
 		nodind4[2] = 2;
@@ -891,7 +891,7 @@ TPZGeoMesh * MalhaGeoT(const int h){//malha triangulo
 		id ++;
 		
 		//-----------------------------------elemento de interface---------------------------
-		TPZVec<int> nodind5(2);
+		TPZVec<long> nodind5(2);
 		
 		nodind5[0]=1;
 		nodind5[1]=4;
@@ -909,7 +909,7 @@ TPZGeoMesh * MalhaGeoT(const int h){//malha triangulo
 		
 		//--------------------------------elementos de contorno----------------------------------------	
 		//omega1
-		TPZVec<int> TopolLine(2);
+		TPZVec<long> TopolLine(2);
 		TopolLine[0] = 0;
 		TopolLine[1] = 1;
 		new TPZGeoElRefPattern<pzgeom:: TPZGeoLinear > (id,TopolLine,mat1BC1,*gmesh);//omega 1
@@ -956,8 +956,8 @@ TPZGeoMesh * MalhaGeoT(const int h){//malha triangulo
 		//	Refinamento uniforme
 		for(int ref = 0; ref < h; ref++){// h indica o numero de refinamentos
 				TPZVec<TPZGeoEl *> filhos;
-				int n = gmesh->NElements();
-				for(int i = 0; i < n; i++){
+				long n = gmesh->NElements();
+				for(long i = 0; i < n; i++){
 						TPZGeoEl * gel = gmesh->ElementVec()[i];
 					
 						if(!gel->HasSubElement() )
